GUI: workaround that prevent Tag (right-click) from crash
This commit disables the ability to pick a new pos from selected
Tag but prevent HeeksCNC from crash under Linux.

This feature is only disable when CMAKE_UNIX is defined, as it seems that
problem does not occurs on Windows (no feedback from MacOS users).

See issue 354 for more details.